Adding Orchestration services SAGA Pattern

+++ registration-service/src/main/java/co/edu/cedesistemas/commerce/registration/services/UserService.java
@@ -2,6 +2,7 @@ package co.edu.cedesistemas.commerce.registration.services;
 
 import co.edu.cedesistemas.commerce.registration.model.User;
 import co.edu.cedesistemas.commerce.registration.repository.UserRepository;
+import co.edu.cedesistemas.common.event.RegistrationEvent;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -11,10 +12,13 @@ import org.springframework.stereotype.Service;
 @Slf4j
 public class UserService {
     private final UserRepository userRepository;
+    private final EventPublisherService publisherService;
 
     public User createUser(User user) {
         user.setStatus(User.Status.INACTIVE);
-        return userRepository.save(user);
+        User usercreated = userRepository.save(user);
+        publisherService.publishRegistrationEvent(usercreated, RegistrationEvent.Status.USER_CREATED);
+        return usercreated;
     }
 
     public User activateUser(String userId){
@@ -27,10 +31,16 @@ public class UserService {
     }
 
     public User getUserById(String userId){
+
         return userRepository.findById(userId).orElse(null);
     }
 
     public void deleteUser(String userId){
-        userRepository.deleteById(userId);
+        User userFound = getUserById(userId);
+        if(userFound != null){
+            userRepository.deleteById(userId);
+            publisherService.publishRegistrationEvent(userFound, RegistrationEvent.Status.USER_DELETED);
+        }
+
     }
 }
